allow loading records from other CHs, CHM-734 (#1363)
* allow loading records from other CHs

* Verify if owner has different name and same env (git suggestions)

* move getOwnerRealm function to realm.js

* remove token

* create validateRealmEnvironment

* move getOwnerRealm to solr api

* Adds getOwnerRealm method to RealmsApi

Implements a method to retrieve the owner realm based on an identifier
using the Solr API. This allows for efficient lookup of realm ownership.

Removes the getOwnerRealm method from the Solr API, consolidating the
functionality into the Realms API.

Updates the Solr API query method to improve readability and parameter
handling.

* Fixes record reference loading in other realms

Ensures record references can be loaded correctly when they
reside in a different realm but the same environment.

It checks if the document exists in the specified realm and if
not, attempts to load it from its owner realm, given that both realms
share the same environment.

This resolves issues where linked records were not being displayed
correctly when moved to a different realm within the same environment.

---------

Co-authored-by: Blaise Fonseca <blaise.fonseca@un.org>