Allow html based application alert chm-810 (#1428)
* allow to pass HTML

* Adds jsdom dependency

Adds jsdom as a project dependency.

jsdom provides a browser-like environment in Node.js, enabling the execution of client-side code in a server-side context.

* Sanitizes site alert to prevent XSS

Ensures the site alert message is sanitized using DOMPurify to mitigate potential cross-site scripting (XSS) vulnerabilities. This enhances the security of the application by preventing the execution of malicious code within the alert message.

---------

Co-authored-by: Blaise Fonseca <blaise.fonseca@un.org>
Co-authored-by: Stephane Bilodeau <stephane.bilodeau@un.org>