Update Matrix, add fields from download api chm-817 (#1427)
* get data from download api

* Get schema name from URL

* Add Generic fields for multiple schemas

* fix fields translated titles

* fix 'Year' field

* get countries code from titles, fix region

* Update Matrix UI, fix the width issues

* fix facets issue

* get facetResult in case of download api

* translate the generic fields

* clean code

* make 'Year' default selected

* remove extra code

* simplify the code logic

* Reset rows and cols to default on schema change

* fix translated country name issue

* feat: Introduce utility function for batching Solr queries

* refactor(matrix): Use batch querying for data loading

* refactor(matrix): Optimize query parameters and schema initialization

* Enhances data matrix for schema-specific downloads

Improves the data matrix directive to fetch records
efficiently using the download API when a specific schema
is selected. This change also includes fallback logic to
use generic fields when no schema or multiple schemas are
selected and adds batch processing to handle large datasets
without overwhelming the browser, improving performance and
user experience.

* Refactors filter query update logic

Moves the filter query string update to before the conditional return.
This ensures the query string is updated even when no filters are selected.

* Replaces console.log with console.error

Updates the logging of errors when fetching country data to use `console.error` instead of `console.log`.
This ensures that errors are properly highlighted and easier to identify in the browser's console.

---------

Co-authored-by: Blaise Fonseca <blaise.fonseca@un.org>